django/requirements.txt
Django==1.11
djangorestframework==3.6.2
demjson==2.2.4
whitenoise==3.3.0
celery==4.0.2
redis==2.10.5
psycopg2==2.7.1
django-celery-beat==1.0.1
requests==2.13.0
gunicorn==19.7.1
feedparser==5.2.1
coverage==4.4